+++ src/test/java/com/zerobase/babdeusilbun/meeting/service/MeetingServiceTest.java
@@ -1,298 +1,303 @@
 package com.zerobase.babdeusilbun.meeting.service;
 
-import static com.zerobase.babdeusilbun.enums.MeetingStatus.*;
+import static com.zerobase.babdeusilbun.enums.MeetingStatus.GATHERING;
+import static com.zerobase.babdeusilbun.enums.MeetingStatus.MEETING_CANCELLED;
 import static com.zerobase.babdeusilbun.enums.PurchaseType.*;
 import static org.assertj.core.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
 
+import com.zerobase.babdeusilbun.domain.Address;
 import com.zerobase.babdeusilbun.domain.Meeting;
 import com.zerobase.babdeusilbun.domain.Purchase;
-import com.zerobase.babdeusilbun.domain.PurchasePayment;
 import com.zerobase.babdeusilbun.domain.Store;
+import com.zerobase.babdeusilbun.domain.User;
 import com.zerobase.babdeusilbun.dto.DeliveryAddressDto;
 import com.zerobase.babdeusilbun.dto.MeetingDto;
 import com.zerobase.babdeusilbun.dto.MetAddressDto;
-import com.zerobase.babdeusilbun.enums.MeetingStatus;
-import com.zerobase.babdeusilbun.enums.PurchaseStatus;
+import com.zerobase.babdeusilbun.enums.PurchaseType;
+import com.zerobase.babdeusilbun.exception.CustomException;
+import com.zerobase.babdeusilbun.exception.ErrorCode;
 import com.zerobase.babdeusilbun.meeting.dto.MeetingRequest;
+import com.zerobase.babdeusilbun.meeting.dto.MeetingRequest.Create;
+import com.zerobase.babdeusilbun.meeting.service.impl.MeetingServiceImpl;
+import com.zerobase.babdeusilbun.repository.MeetingQueryRepository;
 import com.zerobase.babdeusilbun.repository.MeetingRepository;
 import com.zerobase.babdeusilbun.repository.PurchasePaymentRepository;
 import com.zerobase.babdeusilbun.repository.PurchaseRepository;
+import com.zerobase.babdeusilbun.repository.StoreImageRepository;
 import com.zerobase.babdeusilbun.repository.StoreRepository;
+import com.zerobase.babdeusilbun.repository.UserRepository;
 import java.time.LocalDateTime;
-import java.time.Month;
+import java.util.Collections;
 import java.util.List;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
+import java.util.Optional;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.User;
+import org.springframework.data.domain.PageImpl;
+import org.springframework.data.domain.Pageable;
 import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.transaction.annotation.Transactional;
 
-@SpringBootTest
-@ActiveProfiles("test")
-@Transactional
-class MeetingServiceTest {
+@ExtendWith(MockitoExtension.class)
+public class MeetingServiceTest {
 
-  @Autowired
+  @InjectMocks
+  private MeetingServiceImpl meetingService;
+
+  @Mock
   private MeetingRepository meetingRepository;
 
-  @Autowired
-  private MeetingService meetingService;
-  @Autowired
+  @Mock
+  private MeetingQueryRepository meetingQueryRepository;
+
+  @Mock
+  private StoreImageRepository storeImageRepository;
+
+  @Mock
+  private UserRepository userRepository;
+
+  @Mock
   private StoreRepository storeRepository;
-  @Autowired
-  private PurchasePaymentRepository purchasePaymentRepository;
-  @Autowired
+
+  @Mock
   private PurchaseRepository purchaseRepository;
 
-  @BeforeEach
-  void init() {
+  @Mock
+  private PurchasePaymentRepository purchasePaymentRepository;
 
-  }
+  @Mock
+  private UserDetails userDetails;
 
-  @Test
-  @DisplayName("모임 조회 - 결제 마감 시간 순")
-  void searchMeetingByDeadline() {
-
-    PageRequest pageRequest = PageRequest.of(0, 4);
-
-    String searchMenu = "";
-    Long categoryFilter = null;
-
-    Page<MeetingDto> pageable =
-        meetingService.getAllMeetingList(1L, "deadline", searchMenu, categoryFilter, pageRequest);
-
-    List<MeetingDto> content = pageable.getContent();
-
-    assertThat(pageable.getSize()).isEqualTo(4);
-    assertThat(pageable.getTotalElements()).isEqualTo(3);
-    assertThat(pageable.getTotalPages()).isEqualTo(1);
-    assertThat(content.size()).isEqualTo(3);
-
-    assertThat(content.getFirst().getPaymentAvailableAt()).isEqualTo(
-        LocalDateTime.of(2024, Month.AUGUST, 24, 12, 0)
-    );
-    assertThat(content.getLast().getPaymentAvailableAt()).isEqualTo(
-        LocalDateTime.of(2024, Month.AUGUST, 24, 12, 20)
-    );
-  }
+  private User user;
+  private Store store;
+  private Meeting meeting;
 
   @Test
-  @DisplayName("모임 조회 - 배송시간 시간 순")
-  void searchMeetingByShippingTime() {
+  void getAllMeetingList_ShouldReturnPageOfMeetingDtos() {
+    // Given
+    Long schoolId = 1L;
+    String sortCriteria = "name";
+    String searchMenu = "pizza";
+    Long categoryFilter = 1L;
+    Pageable pageable = mock(Pageable.class);
 
-    PageRequest pageRequest = PageRequest.of(0, 4);
-    String searchMenu = "";
-    Long categoryFilter = null;
-
-
-    Page<MeetingDto> pageable =
-        meetingService.getAllMeetingList(1L, "shipping-time", searchMenu, categoryFilter,
-            pageRequest);
-
-    List<MeetingDto> content = pageable.getContent();
-
-    assertThat(pageable.getSize()).isEqualTo(4);
-    assertThat(pageable.getTotalElements()).isEqualTo(3);
-    assertThat(pageable.getTotalPages()).isEqualTo(1);
-    assertThat(content.size()).isEqualTo(3);
-
-    assertThat(content.getFirst().getDeliveredAt()).isEqualTo(
-        LocalDateTime.of(2024, Month.AUGUST, 24, 12, 0)
-    );
-    assertThat(content.getLast().getDeliveredAt()).isEqualTo(
-        LocalDateTime.of(2024, Month.AUGUST, 24, 12, 20)
-    );
-  }
-
-  @Test
-  @DisplayName("모임 조회 - 배달비 순")
-  void searchMeetingByShippingFee() {
-
-    PageRequest pageRequest = PageRequest.of(0, 4);
-    String searchMenu = "";
-    Long categoryFilter = null;
-
-
-    Page<MeetingDto> pageable =
-        meetingService.getAllMeetingList(1L, "shipping-fee", searchMenu, categoryFilter, pageRequest);
-
-    List<MeetingDto> content = pageable.getContent();
-
-    assertThat(pageable.getSize()).isEqualTo(4);
-    assertThat(pageable.getTotalElements()).isEqualTo(3);
-    assertThat(pageable.getTotalPages()).isEqualTo(1);
-    assertThat(content.size()).isEqualTo(3);
-
-    assertThat(content.getFirst().getDeliveryFee()).isEqualTo(3000);
-    assertThat(content.getLast().getDeliveryFee()).isEqualTo(4000);
-  }
-
-  @Test
-  @DisplayName("모임 조회 - 최소 주문금액 순")
-  void searchMeetingByMinPrice() {
-
-    PageRequest pageRequest = PageRequest.of(0, 4);
-    String searchMenu = "";
-    Long categoryFilter = null;
-
-
-    Page<MeetingDto> pageable =
-        meetingService.getAllMeetingList(1L, "min-price", searchMenu, categoryFilter, pageRequest);
-
-    List<MeetingDto> content = pageable.getContent();
-    Long storeAId = content.getLast().getStoreId();
-    Long storeBId = content.getFirst().getStoreId();
-    Store findStoreA = storeRepository.findById(storeBId).get();
-    Store findStoreB = storeRepository.findById(storeAId).get();
-
-    assertThat(pageable.getSize()).isEqualTo(4);
-    assertThat(pageable.getTotalElements()).isEqualTo(3);
-    assertThat(pageable.getTotalPages()).isEqualTo(1);
-    assertThat(content.size()).isEqualTo(3);
-
-    assertThat(findStoreA.getMinOrderAmount()).isEqualTo(1000L);
-    assertThat(findStoreB.getMinOrderAmount()).isEqualTo(2000L);
-  }
-
-  @Test
-  @DisplayName("모임 조회 - 검색어")
-  void searchMeetingBySearchMenu() {
-
-    PageRequest pageRequest = PageRequest.of(0, 4);
-    String searchMenu = "storeB";
-    Long categoryFilter = null;
-
-
-    Page<MeetingDto> pageable =
-        meetingService.getAllMeetingList(1L, "min-price", searchMenu, categoryFilter, pageRequest);
-
-    List<MeetingDto> content = pageable.getContent();
-    Long storeBId = content.getLast().getStoreId();
-    Store findStoreB = storeRepository.findById(storeBId).get();
-
-    assertThat(pageable.getSize()).isEqualTo(4);
-    assertThat(pageable.getTotalElements()).isEqualTo(1);
-    assertThat(pageable.getTotalPages()).isEqualTo(1);
-    assertThat(content.size()).isEqualTo(1);
-
-    assertThat(findStoreB.getName()).isEqualTo("storeB");
-  }
-
-  @Test
-  @DisplayName("모임 정보 조회")
-  void getMeetingInfo() {
-    Long meetingAId = 1L;
-    Long storeBId = 1L;
-    Long userAId = 1L;
-
-    Meeting findMeeting = meetingRepository.findById(meetingAId).get();
-
-    assertThat(findMeeting.getId()).isEqualTo(meetingAId);
-    assertThat(findMeeting.getStore().getId()).isEqualTo(storeBId);
-    assertThat(findMeeting.getLeader().getId()).isEqualTo(userAId);
-
-    assertThat(findMeeting.getPurchaseType()).isEqualTo(DELIVERY_TOGETHER);
-    assertThat(findMeeting.getStatus()).isEqualTo(GATHERING);
-    assertThat(findMeeting.getPaymentAvailableDt())
-        .isEqualTo(LocalDateTime.of(2024, Month.AUGUST, 24, 12, 10));
-  }
-
-  @Test
-  @DisplayName("모임 생성")
-  void createMeeting() {
-
-    UserDetails userDetails = new User("testuser@test.com", "",
-        List.of(new SimpleGrantedAuthority("user")));
-
-    LocalDateTime now = LocalDateTime.now();
-    DeliveryAddressDto deliveryAddressDto = DeliveryAddressDto.builder()
-        .deliveryPostal("dp").deliveryStreetAddress("ds").deliveryDetailAddress("dd").build();
-    MetAddressDto metAddressDto = MetAddressDto.builder()
-        .metPostal("mp").metDetailAddress("md").metStreetAddress("ms").build();
-
-    MeetingRequest.Create request = MeetingRequest.Create.builder()
-        .storeId(1L)
+    User user = User.builder().email("test@example.com").build();
+    Store store = Store.builder().id(1L).name("Test Store").build();
+    Meeting meeting = Meeting.builder()
+        .id(1L)
+        .leader(user)
+        .store(store)
         .purchaseType(DELIVERY_TOGETHER)
-        .minHeadcount(10)
-        .maxHeadcount(20)
+        .minHeadcount(2)
+        .maxHeadcount(5)
         .isEarlyPaymentAvailable(true)
-        .paymentAvailableAt(now)
-        .deliveryAddress(deliveryAddressDto)
-        .metAddress(metAddressDto)
+        .paymentAvailableDt(LocalDateTime.now().plusDays(1))
+        .deliveredAddress(Address.builder().postal("").detailAddress("").streetAddress("").build())
+        .metAddress(Address.builder().postal("").detailAddress("").streetAddress("").build())
+        .status(GATHERING)
+        .build();
+    Page<Meeting> meetings = new PageImpl<>(List.of(meeting));
+
+    when(meetingQueryRepository.findFilteredMeetingList(schoolId, sortCriteria, searchMenu, categoryFilter, pageable))
+        .thenReturn(meetings);
+    when(storeImageRepository.findAllByStoreOrderBySequenceAsc(store))
+        .thenReturn(Collections.emptyList());
+
+    // When
+    Page<MeetingDto> result = meetingService.getAllMeetingList(schoolId, sortCriteria, searchMenu, categoryFilter, pageable);
+
+    // Then
+    assertEquals(1, result.getTotalElements());
+    verify(meetingQueryRepository, times(1)).findFilteredMeetingList(schoolId, sortCriteria, searchMenu, categoryFilter, pageable);
+  }
+
+  @Test
+  void getMeetingInfo_ShouldReturnMeetingDto() {
+    // Given
+    Long meetingId = 1L;
+
+    User user = User.builder().email("test@example.com").build();
+    Store store = Store.builder().id(1L).name("Test Store").build();
+    Meeting meeting = Meeting.builder()
+        .id(1L)
+        .leader(user)
+        .store(store)
+        .purchaseType(PurchaseType.DELIVERY_TOGETHER)
+        .minHeadcount(2)
+        .maxHeadcount(5)
+        .isEarlyPaymentAvailable(true)
+        .paymentAvailableDt(LocalDateTime.now().plusDays(1))
+        .deliveredAddress(Address.builder().postal("").detailAddress("").streetAddress("").build())
+        .metAddress(Address.builder().postal("").detailAddress("").streetAddress("").build())
+        .status(GATHERING)
         .build();
 
-    int preSize = meetingRepository.findAll().size();
+    when(meetingRepository.findById(meetingId)).thenReturn(Optional.of(meeting));
+    when(storeImageRepository.findAllByStoreOrderBySequenceAsc(store)).thenReturn(Collections.emptyList());
+
+    // When
+    MeetingDto result = meetingService.getMeetingInfo(meetingId);
+
+    // Then
+    assertNotNull(result);
+    assertEquals(meetingId, result.getMeetingId());
+    verify(meetingRepository, times(1)).findById(meetingId);
+  }
+
+  @Test
+  void createMeeting_ShouldSaveMeetingAndPurchase() {
+    // Given
+    MeetingRequest.Create request = Create.builder()
+        .storeId(1L)
+        .minHeadcount(1)
+        .maxHeadcount(10)
+        .deliveryAddress(
+            DeliveryAddressDto.builder().deliveryPostal("").deliveryDetailAddress("").deliveryStreetAddress("").build())
+        .metAddress(MetAddressDto.builder().metPostal("").metDetailAddress("").metStreetAddress("").build())
+        .build();
+
+    User user = User.builder().email("test@example.com").build();
+    Store store = Store.builder().id(1L).name("Test Store").build();
+    Meeting meeting = Meeting.builder()
+        .id(1L)
+        .leader(user)
+        .store(store)
+        .purchaseType(PurchaseType.DELIVERY_TOGETHER)
+        .minHeadcount(2)
+        .maxHeadcount(5)
+        .isEarlyPaymentAvailable(true)
+        .paymentAvailableDt(LocalDateTime.now().plusDays(1))
+        .status(GATHERING)
+        .deliveredAddress(Address.builder().postal("").detailAddress("").streetAddress("").build())
+        .metAddress(Address.builder().postal("").detailAddress("").streetAddress("").build())
+        .build();
+
+    when(userDetails.getUsername()).thenReturn(user.getEmail());
+    when(userRepository.findByEmail(user.getEmail())).thenReturn(Optional.of(user));
+    when(storeRepository.findById(any(Long.class))).thenReturn(Optional.of(store));
+    when(meetingRepository.save(any(Meeting.class))).thenReturn(meeting);
+
+    // When
     meetingService.createMeeting(request, userDetails);
-    List<Meeting> findAll = meetingRepository.findAll();
-    Meeting savedMeeting = findAll.getLast();
 
-    assertThat(findAll.size()).isEqualTo(preSize + 1);
-    assertThat(savedMeeting.getLeader().getEmail()).isEqualTo("testuser@test.com");
-    assertThat(savedMeeting.getStore().getId()).isEqualTo(1L);
-    assertThat(savedMeeting.getPurchaseType()).isEqualTo(DELIVERY_TOGETHER);
-    assertThat(savedMeeting.getMinHeadcount()).isEqualTo(10);
-    assertThat(savedMeeting.getMaxHeadcount()).isEqualTo(20);
-    assertThat(savedMeeting.getIsEarlyPaymentAvailable()).isTrue();
-    assertThat(savedMeeting.getPaymentAvailableDt()).isEqualTo(now);
-    assertThat(savedMeeting.getDeliveredAddress().getPostal())
-        .isEqualTo(deliveryAddressDto.getDeliveryPostal());
-    assertThat(savedMeeting.getDeliveredAddress().getStreetAddress())
-        .isEqualTo(deliveryAddressDto.getDeliveryStreetAddress());
-    assertThat(savedMeeting.getDeliveredAddress().getDetailAddress())
-        .isEqualTo(deliveryAddressDto.getDeliveryDetailAddress());
-    assertThat(savedMeeting.getMetAddress().getPostal())
-        .isEqualTo(metAddressDto.getMetPostal());
-    assertThat(savedMeeting.getMetAddress().getStreetAddress())
-        .isEqualTo(metAddressDto.getMetStreetAddress());
-    assertThat(savedMeeting.getMetAddress().getDetailAddress())
-        .isEqualTo(metAddressDto.getMetDetailAddress());
+    // Then
+    verify(meetingRepository, times(1)).save(any(Meeting.class));
+    verify(purchaseRepository, times(1)).save(any(Purchase.class));
   }
 
   @Test
-  @DisplayName("모임 취소 - 리더")
-  void withDrawMeeting_Leader() {
+  void updateMeeting_ShouldUpdateMeetingDetails() {
+    // Given
+    Long meetingId = 1L;
+    MeetingRequest.Update request = MeetingRequest.Update.builder()
+        .maxHeadcount(10)
+        .deliveryAddress(
+            DeliveryAddressDto.builder().deliveryPostal("update").deliveryDetailAddress("update").deliveryStreetAddress("update").build())
+        .metAddress(MetAddressDto.builder().metPostal("update").metDetailAddress("update").metStreetAddress("update").build())
+        .build();
 
-    UserDetails userDetails = new User("testuser@test.com", "",
-        List.of(new SimpleGrantedAuthority("user")));
+    User user = User.builder().email("test@example.com").build();
+    Store store = Store.builder().id(1L).name("Test Store").build();
+    Meeting meeting = Meeting.builder()
+        .id(1L)
+        .leader(user)
+        .store(store)
+        .purchaseType(PurchaseType.DELIVERY_TOGETHER)
+        .minHeadcount(2)
+        .maxHeadcount(5)
+        .isEarlyPaymentAvailable(true)
+        .paymentAvailableDt(LocalDateTime.now().plusDays(1))
+        .status(GATHERING)
+        .build();
 
-    meetingService.withdrawMeeting(1L, userDetails);
-    Meeting findMeeting = meetingRepository.findById(1L).get();
-    Purchase findPurchase = purchaseRepository.findAllByMeeting(findMeeting).getFirst();
-    PurchasePayment findPurchasePayment = purchasePaymentRepository.findLastPurchasePayment(
-        findPurchase.getId()).get();
+    when(userDetails.getUsername()).thenReturn(user.getEmail());
+    when(userRepository.findByEmail(user.getEmail())).thenReturn(Optional.of(user));
+    when(meetingRepository.findById(meetingId)).thenReturn(Optional.of(meeting));
 
-    assertThat(findMeeting.getStatus()).isEqualTo(MEETING_CANCELLED);
-    assertThat(findPurchase.getStatus()).isEqualTo(PurchaseStatus.CANCEL);
-    assertThat(findPurchasePayment).isNotNull();
+    // When
+    meetingService.updateMeeting(meetingId, request, userDetails);
+
+    // Then
+    verify(meetingRepository, times(1)).findById(meetingId);
   }
 
   @Test
-  @DisplayName("모임 취소 - 참가자")
-  void withDrawMeeting_Participant() {
+  void withdrawMeeting_ShouldCancelPurchaseAndUpdateSnapshot() {
+    // Given
+    Long meetingId = 1L;
 
-    UserDetails userDetails = new User("testuser2@test.com", "",
-        List.of(new SimpleGrantedAuthority("user")));
+    User user = User.builder().email("test@example.com").build();
+    Store store = Store.builder().id(1L).name("Test Store").build();
+    Meeting meeting = Meeting.builder()
+        .id(1L)
+        .leader(user)
+        .store(store)
+        .purchaseType(PurchaseType.DELIVERY_TOGETHER)
+        .minHeadcount(2)
+        .maxHeadcount(5)
+        .isEarlyPaymentAvailable(true)
+        .paymentAvailableDt(LocalDateTime.now().plusDays(1))
+        .status(GATHERING)
+        .build();
 
-    meetingService.withdrawMeeting(1L, userDetails);
+    when(userDetails.getUsername()).thenReturn(user.getEmail());
+    when(userRepository.findByEmail(user.getEmail())).thenReturn(Optional.of(user));
+    when(meetingRepository.findById(meetingId)).thenReturn(Optional.of(meeting));
 
-    Meeting findMeeting = meetingRepository.findById(1L).get();
-    Purchase findPurchase = purchaseRepository.findAllByMeeting(findMeeting).getLast();
-    PurchasePayment findPurchasePayment = purchasePaymentRepository.findLastPurchasePayment(
-        findPurchase.getId()).get();
-
-    assertThat(findMeeting.getStatus()).isEqualTo(GATHERING);
-    assertThat(findPurchase.getStatus()).isEqualTo(PurchaseStatus.CANCEL);
-    assertThat(findPurchasePayment).isNotNull();
+    // When
+    // Then
+    CustomException customException = assertThrows(CustomException.class,
+        () -> meetingService.withdrawMeeting(meetingId, userDetails));
+    assertThat(customException.getErrorCode()).isEqualTo(ErrorCode.MEETING_PARTICIPANT_EXIST);
   }
 
+  @Test
+  void withdrawMeeting_ShouldThrowException_WhenMeetingIsNotGathering() {
+    // Given
+    User user = User.builder().email("test@example.com").build();
+    Store store = Store.builder().id(1L).name("Test Store").build();
+    Meeting meeting = Meeting.builder()
+        .id(1L)
+        .leader(user)
+        .store(store)
+        .purchaseType(PurchaseType.DELIVERY_TOGETHER)
+        .minHeadcount(2)
+        .maxHeadcount(5)
+        .isEarlyPaymentAvailable(true)
+        .paymentAvailableDt(LocalDateTime.now().plusDays(1))
+        .status(MEETING_CANCELLED)
+        .build();
 
+    when(userDetails.getUsername()).thenReturn(user.getEmail());
+    when(userRepository.findByEmail(user.getEmail())).thenReturn(Optional.of(user));
+    when(meetingRepository.findById(meeting.getId())).thenReturn(Optional.of(meeting));
+
+    // When & Then
+    CustomException exception = assertThrows(CustomException.class,
+        () -> meetingService.withdrawMeeting(meeting.getId(), userDetails));
+
+    assertThat(exception.getErrorCode()).isEqualTo(ErrorCode.MEETING_STATUS_INVALID);
+
+  }
 
 
 }
+
+
+
+
+
+
+
+
+
